+++ vital_site/vital/views/student_view.py
@@ -57,7 +57,7 @@ def start_novnc(config, started_vm):
         available_config = Available_Config.objects.filter(category='TERM_PORT').order_by('id')[0]
         locked_conf = Available_Config.objects.select_for_update().filter(id=available_config.id)
         cnt += 1
-        if locked_conf is not None:
+        if locked_conf is not None and len(locked_conf) > 0:
             val = locked_conf[0].value
             locked_conf.delete()
             cmd = 'sh /var/www/clone.com/interim/noVNC/utils/launch.sh --listen ' + val + \
